+++ core/src/main/java/org/elasticsearch/rest/action/search/RestClearScrollAction.java
@@ -23,15 +23,12 @@ import org.elasticsearch.action.search.ClearScrollRequest;
 import org.elasticsearch.action.search.ClearScrollResponse;
 import org.elasticsearch.client.node.NodeClient;
 import org.elasticsearch.common.Strings;
-import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.inject.Inject;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.xcontent.XContentHelper;
 import org.elasticsearch.common.xcontent.XContentParser;
 import org.elasticsearch.rest.BaseRestHandler;
 import org.elasticsearch.rest.RestController;
 import org.elasticsearch.rest.RestRequest;
-import org.elasticsearch.rest.action.RestActions;
 import org.elasticsearch.rest.action.RestStatusToXContentListener;
 
 import java.io.IOException;
@@ -56,12 +53,16 @@ public class RestClearScrollAction extends BaseRestHandler {
         clearRequest.setScrollIds(Arrays.asList(splitScrollIds(scrollIds)));
         if (request.hasContentOrSourceParam()) {
             if (request.contentOrSourceParamXContentType() == null) {
-                scrollIds = RestActions.getRestContent(request).utf8ToString();
+                scrollIds = request.contentOrSourceParamString();
                 clearRequest.setScrollIds(Arrays.asList(splitScrollIds(scrollIds)));
             } else {
                 // NOTE: if rest request with xcontent body has request parameters, these parameters does not override xcontent value
                 clearRequest.setScrollIds(null);
-                buildFromContent(RestActions.getRestContent(request), clearRequest);
+                try (XContentParser parser = request.contentOrSourceParamParser()) {
+                    buildFromContent(parser, clearRequest);
+                } catch (IOException e) {
+                    throw new IllegalArgumentException("Failed to parse request body", e);
+                }
             }
         }
 
@@ -75,31 +76,27 @@ public class RestClearScrollAction extends BaseRestHandler {
         return Strings.splitStringByCommaToArray(scrollIds);
     }
 
-    public static void buildFromContent(BytesReference content, ClearScrollRequest clearScrollRequest) {
-        try (XContentParser parser = XContentHelper.createParser(content)) {
-            if (parser.nextToken() != XContentParser.Token.START_OBJECT) {
-                throw new IllegalArgumentException("Malformed content, must start with an object");
-            } else {
-                XContentParser.Token token;
-                String currentFieldName = null;
-                while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
-                    if (token == XContentParser.Token.FIELD_NAME) {
-                        currentFieldName = parser.currentName();
-                    } else if ("scroll_id".equals(currentFieldName) && token == XContentParser.Token.START_ARRAY) {
-                        while ((token = parser.nextToken()) != XContentParser.Token.END_ARRAY) {
-                            if (token.isValue() == false) {
-                                throw new IllegalArgumentException("scroll_id array element should only contain scroll_id");
-                            }
-                            clearScrollRequest.addScrollId(parser.text());
+    public static void buildFromContent(XContentParser parser, ClearScrollRequest clearScrollRequest) throws IOException {
+        if (parser.nextToken() != XContentParser.Token.START_OBJECT) {
+            throw new IllegalArgumentException("Malformed content, must start with an object");
+        } else {
+            XContentParser.Token token;
+            String currentFieldName = null;
+            while ((token = parser.nextToken()) != XContentParser.Token.END_OBJECT) {
+                if (token == XContentParser.Token.FIELD_NAME) {
+                    currentFieldName = parser.currentName();
+                } else if ("scroll_id".equals(currentFieldName) && token == XContentParser.Token.START_ARRAY) {
+                    while ((token = parser.nextToken()) != XContentParser.Token.END_ARRAY) {
+                        if (token.isValue() == false) {
+                            throw new IllegalArgumentException("scroll_id array element should only contain scroll_id");
                         }
-                    } else {
-                        throw new IllegalArgumentException("Unknown parameter [" + currentFieldName
-                                + "] in request body or parameter is of the wrong type[" + token + "] ");
+                        clearScrollRequest.addScrollId(parser.text());
                     }
+                } else {
+                    throw new IllegalArgumentException("Unknown parameter [" + currentFieldName
+                            + "] in request body or parameter is of the wrong type[" + token + "] ");
                 }
             }
-        } catch (IOException e) {
-            throw new IllegalArgumentException("Failed to parse request body", e);
         }
     }
